+++ packages/browser-wallet/src/popup/pages/VerifiableCredentialBackup/VerifiableCredentialImport.tsx
@@ -15,6 +15,7 @@ import { decrypt } from '@shared/utils/crypto';
 import { useHdWallet } from '@popup/shared/utils/account-helpers';
 import JSONBigInt from 'json-bigint';
 import { FileInput } from '@popup/shared/Form/FileInput';
+import { reviveDateFromTimeStampAttribute } from '@concordium/web-sdk';
 import { VerifiableCredentialCardWithStatusFromChain } from '../VerifiableCredential/VerifiableCredentialList';
 import { ExportFormat, VerifiableCredentialExport } from './utils';
 
@@ -46,7 +47,7 @@ async function parseExport(data: EncryptedData, encryptionKey: string): Promise<
     const backup: ExportFormat = JSONBigInt({
         alwaysParseAsBig: true,
         useNativeBigInt: true,
-    }).parse(decrypted);
+    }).parse(decrypted, reviveDateFromTimeStampAttribute);
     // Change index to number, due to parse changing all numbers to bigints.
     backup.value.verifiableCredentials = backup.value.verifiableCredentials.map((v) => ({
         ...v,