+++ packages/browser-wallet/src/popup/pages/Web3ProofRequest/Display/DisplaySecretStatements.tsx
@@ -1,11 +1,10 @@
-import { AttributeType, canProveAtomicStatement, StatementTypes } from '@concordium/web-sdk';
-import { CredentialSchemaSubject } from '@shared/storage/types';
+import { AttributeType, canProveAtomicStatement, CredentialSchemaSubject, StatementTypes } from '@concordium/web-sdk';
 import React from 'react';
 import { TFunction, useTranslation } from 'react-i18next';
 import { DisplayStatementLine } from './DisplayStatementLine';
 import { DisplayBox } from './DisplayBox';
 import { SecretStatementV2 } from '../utils';
-import { DisplayProps, getPropertyTitle } from './utils';
+import { DisplayProps, defaultFormatAttribute, getPropertyTitle } from './utils';
 
 function getStatementValue<Attribute extends AttributeType>(
     statement: SecretStatementV2,
@@ -62,8 +61,8 @@ export function DisplaySecretStatements<Attribute extends AttributeType>({
     statements,
     attributes,
     className,
-    formatAttribute = (_, value) => value.toString(),
     overwriteSecretLine = () => ({}),
+    formatAttribute = defaultFormatAttribute,
 }: DisplayProps<SecretStatementV2, Attribute>) {
     const { t } = useTranslation('web3IdProofRequest', { keyPrefix: 'displayStatement' });
     const header = t('headers.secret');