+++ packages/browser-wallet/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+-   Indicator on proof request page, to show how many credential statements are requested, and the current position.
+
+### Fixed
+
+-   An issue where Date attributes were saved as strings when exported. This would mean that they would lose typing and the credential would be broken.
+-   An issue where statement parameters were not validated according to the attribute bounds.
+
 ## 1.1.4
 
 ### Added